requirements.txt
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cssselect==1.1.0
deprecation==2.1.0
Flask==1.1.2
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpickle==2.0.0
labelary-mwisslead==1.0.1
lxml==4.6.3
MarkupSafe==1.1.1
msgpack==1.0.2
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
parsel==1.6.0
protobuf==3.15.8
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygsheets==2.0.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
pywin32
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
selectorlib==0.16.0
selenium==3.141.0
six==1.15.0
squareup==11.0.0.20210513
uritemplate==3.0.1
urllib3==1.26.6
w3lib==1.22.0
webdriver-manager==3.4.1
Werkzeug==1.0.1
zebra==0.1.0
